+++ stacks/neam/debian-php-nginx.dna-project-base/stack/phphaproxy-stats.sh
@@ -1,0 +1,1 @@
+open $(docker-stack local url phphaproxy 1936 stats:n3v3rg0nnAg1v3y0uup@192.168.99.100)