+++ stacks/neam/debian-php-nginx.dna-project-base/stack/restart.sh
@@ -8,7 +8,7 @@ script_path=`dirname $0`
 cd $script_path/..
 
 # run actual command
-echo "Starting /bin/bash in a new tester container. Press ENTER to see the prompt (some sort of bug...)"
-docker-stack local run tester-$@
+stack/stop.sh
+stack/start.sh
 
 exit 0