+++ stacks/neam/debian-php-nginx.dna-project-base/stack/shell.sh
@@ -9,6 +9,6 @@ cd $script_path/..
 
 # run actual command
 echo "Starting /bin/bash in a new worker container. Press ENTER to see the prompt (some sort of bug...)"
-docker-stack local run worker /bin/bash
+docker-stack local run worker
 
 exit 0