+++ stacks/neam/debian-php-nginx.dna-project-base/stack/src/install-deps.sh
@@ -6,22 +6,10 @@ set -x
 # fail on any error
 set -o errexit
 
-if [ "$PREFER" == "" ]; then
-  PREFER=source
-fi
+export PREFER=$PREFER
 
-composer install --prefer-$PREFER --optimize-autoloader --ignore-platform-reqs
-
-cd dna/
-composer install --prefer-$PREFER --optimize-autoloader --ignore-platform-reqs
-cd -
-
-cd frontend/
-composer install --prefer-$PREFER --optimize-autoloader --ignore-platform-reqs
-cd -
-
-cd backend/
-composer install --prefer-$PREFER --optimize-autoloader --ignore-platform-reqs
-cd -
+script_path=`dirname $0`
+$script_path/install-core-deps.sh
+$script_path/install-local-deps.sh
 
 exit 0